requirements/base.txt
beautifulsoup4==4.6.0
Django==1.11.5
django-extensions==1.9.1
django-taggit==0.22.1
requests==2.18.4
